Ensure all scores are floating points during download
Most scores are floats; ifeval is an exception. The ifeval framework
reports results as Boolean's, or lists of Boolean's. Those values are
reduced to a single number on the HF leaderboard, although it is not
clear how that is done. An assumption is made here to convert the
Boolean's to floats, and in the case of lists take the mean of those
floats.