Ensure that all questions are written the document database
By only writing the document database once, there is a chance that
some questions may be missing: if the respondent for whom the database
is being written did not answer all the questions, for example. This
attempts a write for everyone, keeping track of questions that were
written to avoid duplication.